feat(android) get radius from preferences

+++ raycast-android/app/src/main/java/com/raycast/service/MessageService.java
@@ -53,7 +53,7 @@ public class MessageService extends AbstractCrudService {
     /*
         LIST all messages from API being withing the specified radius of the coordinates
      */
-    public List<Message> list(Location location, double radius){
+    public List<Message> list(Location location, float radius){
         try {
             final String url = contextUrl.buildUpon()
                     .appendQueryParameter("latitude", String.valueOf(location.getLatitude()))
